+++ src/asp/Core/StereoSettings.cc
@@ -412,8 +412,10 @@ FilteringDescription::FilteringDescription(): po::options_description("Filtering
       "Disable filling of holes using an inpainting method. Ignored and obsolete. To be removed in future versions of the software.")
     ("fill-holes-max-size", po::value(&global.fill_hole_max_size)->default_value(100000),
       "Holes with no more pixels than this number should be filled in.")
-    ("edge-buffer-size", po::value(&global.mask_buffer_size)->default_value(-1),
-      "Size of region filtered off the image border.  If unset, equal to the subpixel kernel size.")
+    ("edge-buffer-size", po::value(&global.edge_buffer_size)->default_value(-1),
+      "Crop to be applied around image borders during filtering. If not set, and "
+      "subpixel-mode is between 1 and 6 (inclusive), use the max of two subpixel-kernel "
+      "values. Otherwise set to 5.")
     ("erode-max-size", po::value(&global.erode_max_size)->default_value(0),
       "Isolated blobs with no more pixels than this number should be removed.")
     ("median-filter-size", po::value(&global.median_filter_size)->default_value(0),